+++ CHANGELOG.md
@@ -34,6 +34,7 @@ There is a 3.x to 4.x [migration guide](https://github.com/rmbolger/Posh-ACME/wi
   * This can be disabled per ACME server using a new `DisableTelemetry` parameter in `Set-PAServer`.
   * The data will be used to guide future development decisions in the module.
   * The same User-Agent header is also sent with all calls to the ACME server which is a requirement of the protocol and can't be disabled.
+* Added `NoRefresh` switch to `Set-PAServer` which prevents a request to the ACME server to update endpoint and nonce info. This is useful for updating local preferences without making a server round-trip.
 * BUYPASS_PROD and BUYPASS_TEST are now recognized shortcuts for the the buypass.com CA environments when you use `Set-PAServer`.
 * ZEROSSL_PROD is now a recognized shortcut for the zerossl.com CA when you use `Set-PAServer`.
 * Added tab completion for `DirectoryUrl` in `Set-PAServer`.
@@ -65,6 +66,7 @@ There is a 3.x to 4.x [migration guide](https://github.com/rmbolger/Posh-ACME/wi
 ### Fixes
 
 * Using `Get-PAOrder` with `-Refresh` will no longer throw a terminating error if the ACME server returns an error. It will warn and return the cached copy of the order instead.
+
 
 ## 3.20.0 (2020-11-25)